fixed issue where you could see edit and delete buttons on all comments, fixed issue where likes would not update when changing post page, refactored post page code and divided it into more components, added persisted state and fixed refreshing issues on post, deleted unnecesary dependencies, added post links to profile, fixed issue with paragraph spaces, fixed issue with refreshing profile pages, added avatars
