+++ front-end/package.json
@@ -9,11 +9,9 @@
     "preview": "vite preview"
   },
   "dependencies": {
-    "@taoqf/quill-image-resize-module": "^3.0.1",
-    "@vueup/vue-quill": "^1.2.0",
     "axios": "^1.6.8",
     "pinia": "^2.1.7",
-    "quill": "^2.0.2",
+    "pinia-plugin-persistedstate": "^3.2.1",
     "vue": "^3.4.21",
     "vue-router": "^4.3.2",
     "vuex": "^4.0.2"
