+++ front-end/src/components/ProfileAbout.vue
@@ -1,14 +1,14 @@
 <template>
     <div class="flex w-[100%] h-[40%] bg-black/5">
         <div class="flex flex-col w-[25%] h-[100%] items-center justify-evenly px-16 rounded-lg ">
-            <img src="../assets/avatar-01.png" alt="" class="border-white border-4 rounded-full">
+            <img :src="getAvatarUrl(userProfile.id)" alt="" class="w-[100px] h-[100px] border-white border-4 rounded-full">
             <h1 class="text-2xl font-bold"> {{ userProfile.username }}</h1>
             <p>Joined </p>
             <p>23 Sep 2033</p>
         </div>
         <div class="flex flex-col items-start w-[100%] h-[100%] gap-4 py-6 px-16 ">
             <h2 class="text-xl font-bold">About {{ userProfile.username }}</h2>
-            <div class="flex justify-between w-[80%] text-2xl">
+            <div class="flex justify-around w-[80%] text-2xl">
                 <div class="flex flex-col">
                     <p>Posts</p>
                     <p>{{ posts.length }}</p>
@@ -18,12 +18,13 @@
                     <p>{{ likedPosts.length }}</p>
                 </div>                  
             </div>
-
+            
         </div>
     </div>
 </template>
 <script>
 import axios from '../axios';
+import { getAvatarUrl } from '../utils/getAvatars.js';
 
 export default {
     name: 'ProfileAbout',
@@ -51,7 +52,8 @@ export default {
                 }
             });
             this.likedPosts = response.data;
-        }
+        },
+        getAvatarUrl,
     }
 }
 </script>
