+++ front-end/src/main.js
@@ -3,14 +3,13 @@ import './style.css';
 import App from './App.vue';
 import router from './router/router.js';
 import { createPinia } from 'pinia';
-/* import { QuillEditor } from '@vueup/vue-quill' */
-
-
-const pinia = createPinia();
+import piniaPluginPersistedstate from 'pinia-plugin-persistedstate'
 
 const app = createApp(App);
-/* app.component('QuillEditor', QuillEditor) */
+const pinia = createPinia();
+
 app.use(router);
 app.use(pinia);
+pinia.use(piniaPluginPersistedstate)
 
 app.mount('#app');
