+++ front-end/src/style.css
@@ -31,12 +31,6 @@ body {
 #content :nth-child(5) {
     grid-column: 3 / span 1;
     grid-row: 2 / span 2;
-    background-color: yellow;
+    background-color: #ff9602;
 }
 
-#content :nth-child(3) {
-    background-color: green;
-}
-#content :nth-child(4) {
-    background-color: blue;
-}
