+++ front-end/src/utils/timeAgo.js
@@ -1,0 +1,19 @@
+export function timeAgo(timestamp) {
+    const date = new Date(timestamp);
+    const now = new Date();
+    const diffInMs = now - date;
+    const diffInSeconds = Math.floor(diffInMs / 1000);
+    const diffInMinutes = Math.floor(diffInSeconds / 60);
+    const diffInHours = Math.floor(diffInMinutes / 60);
+    const diffInDays = Math.floor(diffInHours / 24);
+
+    if (diffInSeconds < 60) {
+      return diffInSeconds === 1 ? '1 second ago' : `${diffInSeconds} seconds ago`;
+    } else if (diffInMinutes < 60) {
+      return diffInMinutes === 1 ? '1 minute ago' : `${diffInMinutes} minutes ago`;
+    } else if (diffInHours < 24) {
+      return diffInHours === 1 ? '1 hour ago' : `${diffInHours} hours ago`;
+    } else {
+      return diffInDays === 1 ? '1 day ago' : `${diffInDays} days ago`;
+    }
+  };
